Pull Setter Methods for BingService Into a Trait
