+++ test/Fixtures/ServiceWithoutNamespaceConstant.php
@@ -6,8 +6,5 @@ use PMG\BingAds\BingService;
 
 class ServiceWithoutNamespaceConstant implements BingService
 {
-    use ConverterMethod;
-    use FaultsMethod;
-    use RequestHeadersMethod;
-    use SessionMethod;
+    use \PMG\BingAds\BingServiceDependencies;
 }
